book/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers

from book.models import Book


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Book
        fields = ["id", "title", "description", "created", "updated"]
